refactor: entity 패키지 변경
