+++ src/main/java/subway/repository/dao/LineDao.java
@@ -9,7 +9,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import subway.entity.LineEntity;
+import subway.repository.entity.LineEntity;
 
 @Repository
 public class LineDao {
