+++ src/main/java/subway/repository/dao/SectionDao.java
@@ -7,7 +7,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import subway.entity.SectionEntity;
+import subway.repository.entity.SectionEntity;
 
 @Repository
 public class SectionDao {
