+++ src/main/java/subway/repository/dao/StationDao.java
@@ -10,7 +10,7 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import subway.entity.StationEntity;
+import subway.repository.entity.StationEntity;
 
 @Repository
 public class StationDao {
