+++ src/main/java/subway/repository/entity/SectionEntity.java
@@ -1,4 +1,4 @@
-package subway.entity;
+package subway.repository.entity;
 
 import java.util.Objects;
